Extract dynamic mapping section into its own component. Migrate doc link from schema to documentation service.
